Using json to nest terminal user interface
Sure, here's a very simplified explanation:

This script provides a terminal user interface for executing commands. It uses a JSON file to define a menu structure. Each menu item has an ID, description, and associated command.

- `execute_command(id)`: Executes the command associated with the given ID. If no command is found, it displays an error message.

- `generate_menu(parent_id)`: Generates a menu based on the sub-items of the given parent ID. The user can select an option from this menu.

The script uses `whiptail` to display the menu and `jq` to parse the JSON file.